修改

+++ dataflow/MongodbSync/OSSPusher.py
@@ -4,7 +4,7 @@ import time
 import oss2
 import logging
 from logging.handlers import RotatingFileHandler
-
+import traceback
 
 # Redis 信息
 REDIS_IP = '47.97.27.84'
@@ -48,15 +48,17 @@ class OSSPusher(threading.Thread):
                 oss_new = oss_data.replace('oss-cn-hangzhou', 'oss-cn-hongkong')
                 file_name = oss_data.split('aliyuncs.com/')[-1]
                 try:
-                    if not self.bucket_hk.object_exists(file_name):
+                    if not self.bucket_hk.object_exists(file_name) and self.bucket_hz.object_exists(file_name):
                         # print('开始下载', oss_data)
                         file_stream = self.bucket_hz.get_object(file_name)
                         # print('开始上传', oss_new)
                         self.bucket_hk.put_object(file_name, file_stream)
                         self.logger.info(str(r.llen(OSS_QUEUE)) + '    转写 oss 成功，oss 为: ' + oss_new)
                 except Exception as e:
-                    self.logger.error(str(r.llen(OSS_QUEUE)) + '    转写 oss 失败，oss 为: ' + oss_new + '错误为: ' + str(e))
+                    self.logger.error(str(r.llen(OSS_QUEUE)) + '    转写 oss 失败，oss 为: ' + oss_new + '错误为: \n'
+                                      + traceback.format_exc())
                     r.rpush(OSS_QUEUE, oss_data)
+                    time.sleep(0.001)
             else:
                 self.logger.info('Redis oss 队列中无数据，等待10s再取')
                 time.sleep(10)
